experimento-2/auth/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
gunicorn==21.2.0